pyproject.toml
[project]
name = "getmebuddy"
version = "0.1.0"
description = "Social buddy matching platform"
authors = [
    {name = "Vivek Singh viki00762@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"


[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
python-dotenv = "^1.0.0"
channels = "^4.0.0"
channels-redis = "^4.1.0"
daphne = "^4.0.0"
firebase-admin = "^6.2.0"
django-storages = "^1.14.2"
boto3 = "^1.28.78"
pillow = "^10.1.0"
django-cors-headers = "^4.3.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
sentry-sdk = "^1.34.0"
django-filter = "^23.3"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"
django-debug-toolbar = "^4.2.0"
faker = "^19.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "getmebuddy.settings"
python_files = "test_*.py"
testpaths = ["users", "profiles", "matching", "messaging", "engagement", "safety", "monetization"]